+++ test/integration/bigip/property/propertiesTlsClient.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2023 F5, Inc.
+ * Copyright 2024 F5, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -218,6 +218,11 @@ describe('TLS_Client', function () {
                 name: 'sslSignHash',
                 inputValue: [undefined, 'sha256', undefined],
                 expectedValue: ['any', 'sha256', 'any']
+            },
+            {
+                name: 'handshakeTimeout',
+                inputValue: [undefined, 1234, undefined],
+                expectedValue: [10, 1234, 10]
             }
         ];